+++ demo/src/examples/gif-default.js
@@ -1,17 +1,13 @@
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import Image from 'gatsby-image';
 import { GatsbyImage } from 'gatsby-plugin-image';
 
 const GifDefault = () => {
   const data = useStaticQuery(graphql`
     query {
-      image: file(name: { eq: "victory" }) {
+      image: file(name: { eq: "giphyCat" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(layout: CONSTRAINED, placeholder: BLURRED)
-          fluid {
-            ...CloudinaryAssetFluid
-          }
         }
       }
     }
@@ -20,12 +16,9 @@ const GifDefault = () => {
   // Duplicate the query so we can display it on the page.
   const query = `
     query {
-      image: file(name: { eq: "victory" }) {
+      image: file(name: { eq: "giphyCat" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(layout: CONSTRAINED, placeholder: BLURRED)
-          fluid {
-            ...CloudinaryAssetFluid
-          }
         }
       }
     }
@@ -37,14 +30,10 @@ const GifDefault = () => {
     <div className="image-example">
       <h2>Lazy-loaded animated GIFs!</h2>
 
-      <h3>gatsby-plugin-image</h3>
       <GatsbyImage
         image={data.image.cloudinary.gatsbyImageData}
-        alt="Jason, victorious."
+        alt="Cat Pirate GIF By Product Hunt."
       />
-
-      <h3>gatsby-image</h3>
-      <Image fluid={data.image.cloudinary.fluid} alt="Jason, victorious." />
 
       <h3>Query</h3>
       <pre>{query}</pre>