+++ demo/src/examples/remote-image-1.js
@@ -1,22 +1,19 @@
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import Image from 'gatsby-image';
 import { GatsbyImage } from 'gatsby-plugin-image';
 
 const RemoteImage1 = () => {
   const data = useStaticQuery(graphql`
     query {
-      remoteExample(name: { eq: "Remote Example 1" }) {
-        cloudinary: remoteImage {
+      project(name: { eq: "Project Example One" }) {
+        coverImage {
           gatsbyImageData(
             layout: FIXED
-            height: 600
+            height: 300
+            aspectRatio: 1
             placeholder: TRACED_SVG
-            chained: ["t_lwj"]
+            transformations: ["c_crop", "e_pixelate_faces"]
           )
-          fixed(height: 600, chained: ["t_lwj"]) {
-            ...CloudinaryAssetFixed
-          }
         }
       }
     }
@@ -25,17 +22,15 @@ const RemoteImage1 = () => {
   // Duplicate the query so we can display it on the page.
   const query = `
     query {
-      remoteExample(name: {eq: "Remote Example 1"}) {
-        cloudinary: remoteImage {
+      project(name: { eq: "Project Example One" }) {
+        coverImage {
           gatsbyImageData(
             layout: FIXED
-            height: 600
+            height: 300
+            aspectRatio: 1
             placeholder: TRACED_SVG
-            chained: ["t_lwj"]
+            transformations: ["c_crop", "e_pixelate_faces"]
           )
-          fixed(height: 600, chained: ["t_lwj"]) {
-            ...CloudinaryAssetFixed
-          }
         }
       }
     }
@@ -45,16 +40,12 @@ const RemoteImage1 = () => {
 
   return (
     <div className="image-example">
-      <h2>Example 1</h2>
+      <h2>Remote image</h2>
 
-      <h3>gatsby-plugin-image</h3>
       <GatsbyImage
-        image={data.remoteExample.cloudinary.gatsbyImageData}
-        alt="Jason, victorious."
+        image={data.project.coverImage.gatsbyImageData}
+        alt="Drone photo of a Pirate ship with a US flag by Austin Neill from Unsplash."
       />
-
-      <h3>gatsby-image</h3>
-      <Image fixed={data.remoteExample.cloudinary.fixed} alt="Remote image" />
 
       <h3>Query</h3>
       <pre>{query}</pre>