+++ demo/src/examples/uploaded-chaining.js
@@ -2,15 +2,19 @@ import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
 import { GatsbyImage } from 'gatsby-plugin-image';
 
-const GifTransformed = () => {
+const UploadedChaining = () => {
   const data = useStaticQuery(graphql`
     query {
-      image: file(name: { eq: "giphyCat" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(
-            layout: CONSTRAINED
-            placeholder: BLURRED
-            transformations: ["e_gamma:500"]
+            height: 300
+            layout: FIXED
+            transformations: ["t_gatsby-demo"]
+            chained: [
+              "co_rgb:FFFF00,l_text:Times_90_bold:Cool%2520text"
+              "fl_layer_apply,g_south,y_20"
+            ]
           )
         }
       }
@@ -20,12 +24,16 @@ const GifTransformed = () => {
   // Duplicate the query so we can display it on the page.
   const query = `
     query {
-      image: file(name: { eq: "giphyCat" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(
-            layout: CONSTRAINED
-            placeholder: BLURRED
-            transformations: ["e_gamma:500"]
+            height: 300
+            layout: FIXED
+            transformations: ["t_gatsby-demo"]
+            chained: [
+              "co_rgb:FFFF00,l_text:Times_90_bold:Cool%2520text"
+              "fl_layer_apply,g_south,y_20"
+            ]
           )
         }
       }
@@ -36,11 +44,11 @@ const GifTransformed = () => {
 
   return (
     <div className="image-example">
-      <h2>GIFs with transformations!</h2>
+      <h2>For complex effects, use chaining</h2>
 
       <GatsbyImage
         image={data.image.cloudinary.gatsbyImageData}
-        alt="Cat Pirate GIF By Product Hunt."
+        alt="Pirate photo by Sergey Semin from Unsplash."
       />
 
       <h3>Query</h3>
@@ -49,4 +57,4 @@ const GifTransformed = () => {
   );
 };
 
-export default GifTransformed;
+export default UploadedChaining;