+++ demo/src/examples/uploaded-default.js
@@ -1,17 +1,14 @@
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import Image from 'gatsby-image';
+
 import { GatsbyImage } from 'gatsby-plugin-image';
 
-const FluidDefault = () => {
+const UploadedDefault = () => {
   const data = useStaticQuery(graphql`
     query {
-      image: file(name: { eq: "marisa" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
-          gatsbyImageData(layout: CONSTRAINED)
-          fluid {
-            ...CloudinaryAssetFluid
-          }
+          gatsbyImageData(height: 300, layout: FIXED)
         }
       }
     }
@@ -20,12 +17,9 @@ const FluidDefault = () => {
   // Duplicate the query so we can display it on the page.
   const query = `
     query {
-      image: file(name: { eq: "marisa" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
-          gatsbyImageData(layout: CONSTRAINED)
-          fluid {
-            ...CloudinaryAssetFluid
-          }
+          gatsbyImageData(height: 300, layout: FIXED)
         }
       }
     }
@@ -35,18 +29,11 @@ const FluidDefault = () => {
 
   return (
     <div className="image-example">
-      <h2>Fluid images loaded from Cloudinary</h2>
+      <h2>Straight forward demo</h2>
 
-      <h3>gatsby-plugin-image</h3>
       <GatsbyImage
         image={data.image.cloudinary.gatsbyImageData}
-        alt="Marisa Morby standing in a rose garden."
-      />
-
-      <h3>gatsby-image</h3>
-      <Image
-        fluid={data.image.cloudinary.fluid}
-        alt="Marisa Morby standing in a rose garden."
+        alt="Pirate photo by Sergey Semin from Unsplash."
       />
 
       <h3>Query</h3>
@@ -55,4 +42,4 @@ const FluidDefault = () => {
   );
 };
 
-export default FluidDefault;
+export default UploadedDefault;