+++ demo/src/examples/uploaded-thumb.js
@@ -1,22 +1,18 @@
 import React from 'react';
 import { graphql, useStaticQuery } from 'gatsby';
-import Image from 'gatsby-image';
 import { GatsbyImage } from 'gatsby-plugin-image';
 
-const FixedThumb = () => {
+const UploadedThumb = () => {
   const data = useStaticQuery(graphql`
     query {
-      image: file(name: { eq: "jason" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(
-            width: 225
-            aspectRatio: 1
+            height: 300
+            aspectRatio: 0.8
             layout: FIXED
             transformations: ["c_thumb", "g_face"]
           )
-          fixed(width: 225, transformations: ["ar_1:1", "c_thumb", "g_face"]) {
-            ...CloudinaryAssetFixed
-          }
         }
       }
     }
@@ -25,20 +21,14 @@ const FixedThumb = () => {
   // Duplicate the query so we can display it on the page.
   const query = `
     query {
-      image: file(name: { eq: "jason" }) {
+      image: file(name: { eq: "sergey-semin-unsplash" }) {
         cloudinary: childCloudinaryAsset {
           gatsbyImageData(
-            width: 225
-            aspectRatio: 1
+            height: 300
+            aspectRatio: 0.8
             layout: FIXED
             transformations: ["c_thumb", "g_face"]
           )
-          fixed(
-            width: 225,
-            transformations: ["ar_1:1", "c_thumb", "g_face"]
-          ) {
-            ...CloudinaryAssetFixed
-          }
         }
       }
     }
@@ -48,18 +38,11 @@ const FixedThumb = () => {
 
   return (
     <div className="image-example">
-      <h2>Create avatars from any image!</h2>
+      <h2>Create avatars from any image</h2>
 
-      <h3>gatsby-plugin-image</h3>
       <GatsbyImage
         image={data.image.cloudinary.gatsbyImageData}
-        alt="Jason giving finger guns toward the camera."
-      />
-
-      <h3>gatsby-image</h3>
-      <Image
-        fixed={data.image.cloudinary.fixed}
-        alt="Jason giving finger guns toward the camera."
+        alt="Pirate photo by Sergey Semin from Unsplash."
       />
 
       <h3>Query</h3>
@@ -68,4 +51,4 @@ const FixedThumb = () => {
   );
 };
 
-export default FixedThumb;
+export default UploadedThumb;