+++ demo/src/pages/index.js
@@ -1,9 +1,10 @@
 import React from 'react';
 import Layout from '../components/layout';
-import FluidDefault from '../examples/fluid-default';
-import FixedThumb from '../examples/fixed-thumb';
 import GifDefault from '../examples/gif-default';
-import FluidSilly from '../examples/fluid-silly';
+import UploadedDefault from '../examples/uploaded-default';
+import UploadedThumb from '../examples/uploaded-thumb';
+import ExistingData from '../examples/existing-data';
+import RemoteImage1 from '../examples/remote-image-1';
 
 const IndexPage = () => (
   <Layout>
@@ -12,13 +13,13 @@ const IndexPage = () => (
       <a href="https://cloudinary.com">Cloudinary</a> is really good at managing
       assets. <a href="https://gatsbyjs.org">Gatsby</a> is really good at
       delivering high-performance web pages using (among other things){' '}
-      <a href="https://www.gatsbyjs.org/packages/gatsby-image/">
-        <code>gatsby-image</code>
+      <a href="https://www.gatsbyjs.com/plugins/gatsby-plugin-image/">
+        <code>gatsby-plugin-image</code>
       </a>
       .
     </p>
     <p>
-      Put ’em together and they’re like peanut butter and jelly. Peas and
+      Put’em together and they’re like peanut butter and jelly. Peas and
       carrots. Chocolate and everything.
     </p>
     <p>
@@ -28,9 +29,9 @@ const IndexPage = () => (
     </p>
     <p>
       Under the hood, your images will be uploaded to Cloudinary and added back
-      to Gatsby’s data layer. No more dealing with Sharp issues! No more waiting
-      for hundreds of images to be generated on your local machine! Just sweet,
-      sweet, high-performance images with a{' '}
+      to the Gatsby’s data layer. No more dealing with Sharp issues! No more
+      waiting for hundreds of images to be generated on your local machine! Just
+      sweet, sweet, high-performance images with a{' '}
       <a href="https://cloudinary.com/documentation/image_transformation_reference">
         powerful image transformation API
       </a>{' '}
@@ -43,10 +44,11 @@ const IndexPage = () => (
       </a>
     </p>
     <div className="examples">
-      <FluidDefault />
-      <FixedThumb />
+      <UploadedDefault />
+      <UploadedThumb />
+      <ExistingData variant="minimal" />
+      <RemoteImage1 />
       <GifDefault />
-      <FluidSilly />
     </div>
   </Layout>
 );