+++ plugin/gatsby-node.js
@@ -1,33 +1,15 @@
-const { setPluginOptions, getPluginOptions } = require('./options');
+const {
+  initializaGlobalState,
+  getCoreSupportsOnPluginInit,
+} = require('./options');
 const {
   createCloudinaryAssetType,
   createCloudinaryAssetNodes,
 } = require('./node-creation');
-const {
-  createGatsbyImageResolvers,
-  addGatsbyImageFragments,
-  createGatsbyImageTypes,
-} = require('./gatsby-image');
+
 const { createGatsbyImageDataResolver } = require('./gatsby-plugin-image');
 
-let coreSupportsOnPluginInit = undefined;
-
-try {
-  const { isGatsbyNodeLifecycleSupported } = require(`gatsby-plugin-utils`);
-  if (isGatsbyNodeLifecycleSupported(`onPluginInit`)) {
-    coreSupportsOnPluginInit = 'stable';
-  } else if (isGatsbyNodeLifecycleSupported(`unstable_onPluginInit`)) {
-    coreSupportsOnPluginInit = 'unstable';
-  }
-} catch (error) {
-  console.error(
-    `[gatsby-transformer-cloudinary] Cannot check if Gatsby supports onPluginInit`
-  );
-}
-
-const initializaGlobalState = ({ reporter }, pluginOptions) => {
-  setPluginOptions({ reporter, pluginOptions });
-};
+let coreSupportsOnPluginInit = getCoreSupportsOnPluginInit();
 
 if (coreSupportsOnPluginInit === 'stable') {
   exports.onPluginInit = initializaGlobalState;
@@ -37,28 +19,34 @@ if (coreSupportsOnPluginInit === 'stable') {
   exports.onPreInit = initializaGlobalState;
 }
 
-exports.onPreExtractQueries = async (gatsbyUtils) => {
-  // Fragments to be used with gatsby-image
-  await addGatsbyImageFragments(gatsbyUtils);
+exports.pluginOptionsSchema = ({ Joi }) => {
+  return Joi.object({
+    cloudName: Joi.string(),
+    apiKey: Joi.string(),
+    apiSecret: Joi.string(),
+    uploadFolder: Joi.string(),
+    uploadSourceInstanceNames: Joi.array().items(Joi.string()),
+    transformTypes: Joi.array()
+      .items(Joi.string())
+      .default(['CloudinaryAsset']),
+    overwriteExisting: Joi.boolean().default(false),
+    defaultTransformations: Joi.array()
+      .items(Joi.string())
+      .default(['c_fill', 'g_auto', 'q_auto']),
+  });
 };
 
 exports.createSchemaCustomization = (gatsbyUtils) => {
   // Type to be used for node creation
   createCloudinaryAssetType(gatsbyUtils);
-
-  // Types to be used with gatsby-image
-  createGatsbyImageTypes(gatsbyUtils);
 };
 
-exports.createResolvers = (gatsbyUtils) => {
-  // Resolvers to be used with gatsby-image
-  createGatsbyImageResolvers(gatsbyUtils, getPluginOptions());
-
+exports.createResolvers = (gatsbyUtils, pluginOptions) => {
   // Resolvers to be used with gatsby-plugin-image
-  createGatsbyImageDataResolver(gatsbyUtils, getPluginOptions());
+  createGatsbyImageDataResolver(gatsbyUtils, pluginOptions);
 };
 
-exports.onCreateNode = async (gatsbyUtils) => {
-  // Create Cloudinary Asset nodes if applicable
-  await createCloudinaryAssetNodes(gatsbyUtils, getPluginOptions());
+exports.onCreateNode = async (gatsbyUtils, pluginOptions) => {
+  // Upload and create Cloudinary Asset nodes if applicable
+  await createCloudinaryAssetNodes(gatsbyUtils, pluginOptions);
 };